fixed polling for pipeline and stackComponents
